pyproject.toml
[tool.poetry]
name = "hearth-and-kin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.0"
setuptools = "^69.0.2"
bcrypt = "^4.1.1"
elevenlabs = "0.2.27"
openai = "^1.3.6"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"
langchain = ">0.0.343"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlmodel = "^0.0.14"
pymongo = "^4.6.1"
motor = "^3.3.2"
httpx = "^0.27.0"
langchain-groq = "^0.0.1"
langchain-openai = "^0.0.8"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.21.1"
pytest-dependency = "^0.5.1"
pytest-mock = "^3.12.0"
pytest = "^7.4.3"
ruff = "^0.1.6"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

target-version = "py312"

[tool.ruff.lint]
ignore = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"